requirements.txt
django
djangorestframework
environs
coverage
djangorestframework-simplejwt
django-cors-headers
python-dotenv
mysqlclient
django-filter